Do not remove broadcast dir on resume

+++ src/prime_rl/rl.py
@@ -465,8 +465,9 @@ def rl(config: RLConfig):
         log_dir.mkdir(parents=True, exist_ok=True)
 
         # Cleaning broadcast dir (so that orchestrator does not pre-maturely update weights)
-        logger.info(f"Cleaning broadcast directory ({broadcast_dir})")
-        shutil.rmtree(broadcast_dir, ignore_errors=True)
+        if not (config.ckpt and config.ckpt.resume_step and config.trainer.weight_broadcast == "filesystem"):
+            logger.info(f"Cleaning broadcast directory ({broadcast_dir})")
+            shutil.rmtree(broadcast_dir, ignore_errors=True)
 
         # Cleaning rollouts (so that trainer does not train on old rollouts)
         logger.info(f"Cleaning rollout dir ({rollout_dir})")
